Move statsd filters into their own filter objects.
