+++ lib/logging/transports/statsd.js
@@ -6,6 +6,8 @@ const util = require('util');
 const winston = require('winston');
 const logger = require('../logging').logger;
 const config = require('../../configuration');
+const timing_filter = require('./filters/statsd-timing');
+const increment_filter = require('./filters/statsd-increment');
 
 var StatsD = false;
 try {
@@ -13,8 +15,6 @@ try {
 } catch (requireError) {
   // its ok, its an optionalDependency
 }
-
-const PREFIX = "browserid." + config.get('process_type') + ".";
 
 "use strict";
 
@@ -31,11 +31,11 @@ util.inherits(StatsdTransport, winston.Transport);
 StatsdTransport.prototype.log = function(level, msg, meta, callback) {
   if ( ! this.statsd) return callback(null, true);
 
-  if (isIncrementMessage(msg)) {
-    this.statsd.increment(getIncrementCounterName(msg), meta);
+  if (increment_filter.test(msg)) {
+    this.statsd.increment(increment_filter.toType(msg), meta);
   }
-  else if (isTimingMessage(msg)) {
-    this.statsd.timing(getTimingCounterName(msg), meta, meta);
+  else if (timing_filter.test(msg)) {
+    this.statsd.timing(timing_filter.toType(msg), meta, meta);
   }
 
   callback(null, true);
@@ -56,55 +56,6 @@ function getStatsdIfEnabled() {
 
     return new StatsD(statsdOptions.host, statsdOptions.port);
   }
-}
-
-
-var IncrementMessageMatches = {
-  "assertion_failure": true,
-  "uncaught_exception": true
-};
-
-var IncrementRegExpMatches = [
-  /^wsapi_code_mismatch\./,
-  /^wsapi\./
-];
-
-function isIncrementMessage(msg) {
-  if (msg in IncrementMessageMatches) return true;
-
-  for (var i = 0, regExp; regExp = IncrementRegExpMatches[i]; ++i) {
-    if (regExp.test(msg)) return true;
-  }
-
-  return false;
-}
-
-function getIncrementCounterName(msg) {
-  return PREFIX + msg;
-}
-
-var TimingMessageMatches = {
-  "bcrypt.compare_time": true,
-  "query_time": true,
-  "certification_time": true,
-  "assertion_verification_time": true
-};
-
-var TimingRegExpMatches = [
-];
-
-function isTimingMessage(msg) {
-  if (msg in TimingMessageMatches) return true;
-
-  for (var i = 0, regExp; regExp = TimingRegExpMatches[i]; ++i) {
-    if (regExp.test(msg)) return true;
-  }
-
-  return false;
-}
-
-function getTimingCounterName(msg) {
-  return PREFIX + msg;
 }
